requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-migrate==4.0.5
python-dotenv==1.0.0
groq==0.4.1
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
bcrypt==4.1.1
pyopenssl
